\name{eigenMapMatMult2}
\alias{eigenMapMatMult2}
\title{
eigenMapMatMult2
}
\description{
Matrix multiplication in C++ between two matrices.
}
\usage{
eigenMapMatMult2(A, B)
}
\arguments{
  \item{A}{
First Matrix
}
  \item{B}{
Second Matrix
}
}
\value{
Returns the matrix multiplication of A*B
}
