\name{eigenMapMatMult3}
\alias{eigenMapMatMult3}
\title{
eigenMapMatMult3
}
\description{
Matrix multiplication between 3 matrices (A * B * C)
}
\usage{
eigenMapMatMult3(A, B, C)
}
\arguments{
  \item{A}{
First Matrix
}
  \item{B}{
Second Matrix
}
  \item{C}{
Third Matrix
}
}
\value{
The result of A * B * C
}
