% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BICOSS.R
\docType{data}
\name{SNPs}
\alias{SNPs}
\title{A. Thaliana Genotype matrix}
\format{
## `SNPs`
A matrix with 328 observations and 9,000 SNPs.
}
\usage{
SNPs
}
\description{
This is a matrix with 328 observations and 9,000 SNPs. Each row is contains 9,000 SNPs from a single A. Thaliana ecotype in the paper
Components of Root Architecture Remodeling in Response to Salt Stress.
}
\keyword{datasets}
