\name{vignette_kinship_dat}
\alias{vignette_kinship_dat}
\docType{data}
\title{
vignette_kinship_dat
}
\description{
Dataset associated with the Vignette for the GWAS.BAYES package.
}
\usage{data("vignette_kinship_dat")}
\format{
  A data frame with 4075 observations on the following 1001 variables.
  \describe{
    \item{\code{Phenotype}}{a numeric vector}
    \item{\code{SNP1}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP2}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP3}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP4}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP5}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP6}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP7}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP8}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP9}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP10}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP11}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP12}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP13}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP14}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP15}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP16}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP17}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP18}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP19}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP20}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP21}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP22}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP23}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP24}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP25}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP26}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP27}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP28}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP29}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP30}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP31}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP32}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP33}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP34}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP35}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP36}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP37}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP38}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP39}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP40}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP41}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP42}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP43}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP44}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP45}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP46}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP47}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP48}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP49}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP50}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP51}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP52}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP53}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP54}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP55}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP56}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP57}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP58}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP59}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP60}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP61}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP62}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP63}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP64}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP65}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP66}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP67}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP68}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP69}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP70}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP71}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP72}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP73}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP74}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP75}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP76}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP77}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP78}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP79}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP80}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP81}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP82}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP83}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP84}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP85}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP86}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP87}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP88}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP89}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP90}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP91}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP92}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP93}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP94}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP95}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP96}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP97}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP98}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP99}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP100}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP101}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP102}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP103}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP104}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP105}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP106}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP107}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP108}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP109}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP110}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP111}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP112}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP113}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP114}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP115}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP116}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP117}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP118}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP119}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP120}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP121}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP122}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP123}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP124}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP125}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP126}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP127}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP128}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP129}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP130}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP131}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP132}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP133}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP134}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP135}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP136}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP137}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP138}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP139}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP140}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP141}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP142}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP143}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP144}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP145}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP146}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP147}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP148}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP149}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP150}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP151}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP152}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP153}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP154}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP155}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP156}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP157}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP158}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP159}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP160}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP161}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP162}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP163}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP164}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP165}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP166}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP167}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP168}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP169}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP170}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP171}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP172}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP173}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP174}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP175}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP176}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP177}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP178}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP179}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP180}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP181}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP182}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP183}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP184}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP185}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP186}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP187}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP188}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP189}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP190}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP191}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP192}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP193}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP194}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP195}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP196}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP197}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP198}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP199}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP200}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP201}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP202}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP203}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP204}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP205}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP206}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP207}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP208}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP209}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP210}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP211}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP212}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP213}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP214}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP215}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP216}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP217}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP218}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP219}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP220}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP221}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP222}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP223}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP224}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP225}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP226}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP227}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP228}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP229}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP230}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP231}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP232}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP233}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP234}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP235}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP236}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP237}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP238}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP239}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP240}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP241}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP242}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP243}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP244}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP245}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP246}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP247}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP248}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP249}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP250}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP251}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP252}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP253}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP254}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP255}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP256}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP257}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP258}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP259}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP260}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP261}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP262}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP263}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP264}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP265}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP266}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP267}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP268}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP269}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP270}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP271}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP272}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP273}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP274}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP275}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP276}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP277}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP278}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP279}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP280}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP281}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP282}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP283}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP284}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP285}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP286}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP287}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP288}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP289}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP290}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP291}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP292}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP293}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP294}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP295}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP296}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP297}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP298}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP299}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP300}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP301}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP302}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP303}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP304}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP305}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP306}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP307}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP308}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP309}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP310}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP311}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP312}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP313}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP314}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP315}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP316}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP317}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP318}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP319}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP320}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP321}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP322}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP323}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP324}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP325}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP326}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP327}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP328}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP329}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP330}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP331}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP332}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP333}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP334}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP335}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP336}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP337}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP338}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP339}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP340}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP341}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP342}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP343}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP344}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP345}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP346}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP347}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP348}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP349}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP350}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP351}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP352}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP353}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP354}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP355}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP356}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP357}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP358}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP359}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP360}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP361}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP362}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP363}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP364}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP365}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP366}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP367}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP368}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP369}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP370}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP371}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP372}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP373}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP374}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP375}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP376}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP377}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP378}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP379}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP380}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP381}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP382}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP383}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP384}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP385}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP386}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP387}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP388}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP389}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP390}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP391}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP392}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP393}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP394}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP395}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP396}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP397}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP398}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP399}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP400}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP401}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP402}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP403}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP404}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP405}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP406}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP407}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP408}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP409}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP410}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP411}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP412}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP413}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP414}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP415}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP416}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP417}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP418}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP419}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP420}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP421}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP422}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP423}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP424}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP425}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP426}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP427}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP428}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP429}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP430}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP431}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP432}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP433}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP434}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP435}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP436}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP437}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP438}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP439}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP440}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP441}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP442}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP443}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP444}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP445}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP446}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP447}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP448}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP449}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP450}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP451}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP452}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP453}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP454}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP455}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP456}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP457}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP458}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP459}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP460}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP461}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP462}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP463}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP464}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP465}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP466}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP467}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP468}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP469}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP470}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP471}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP472}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP473}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP474}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP475}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP476}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP477}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP478}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP479}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP480}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP481}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP482}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP483}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP484}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP485}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP486}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP487}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP488}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP489}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP490}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP491}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP492}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP493}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP494}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP495}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP496}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP497}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP498}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP499}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP500}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP501}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP502}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP503}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP504}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP505}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP506}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP507}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP508}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP509}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP510}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP511}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP512}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP513}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP514}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP515}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP516}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP517}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP518}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP519}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP520}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP521}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP522}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP523}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP524}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP525}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP526}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP527}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP528}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP529}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP530}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP531}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP532}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP533}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP534}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP535}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP536}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP537}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP538}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP539}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP540}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP541}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP542}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP543}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP544}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP545}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP546}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP547}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP548}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP549}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP550}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP551}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP552}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP553}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP554}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP555}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP556}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP557}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP558}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP559}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP560}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP561}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP562}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP563}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP564}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP565}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP566}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP567}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP568}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP569}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP570}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP571}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP572}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP573}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP574}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP575}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP576}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP577}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP578}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP579}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP580}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP581}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP582}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP583}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP584}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP585}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP586}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP587}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP588}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP589}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP590}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP591}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP592}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP593}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP594}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP595}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP596}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP597}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP598}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP599}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP600}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP601}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP602}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP603}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP604}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP605}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP606}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP607}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP608}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP609}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP610}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP611}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP612}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP613}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP614}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP615}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP616}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP617}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP618}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP619}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP620}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP621}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP622}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP623}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP624}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP625}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP626}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP627}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP628}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP629}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP630}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP631}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP632}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP633}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP634}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP635}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP636}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP637}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP638}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP639}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP640}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP641}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP642}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP643}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP644}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP645}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP646}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP647}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP648}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP649}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP650}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP651}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP652}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP653}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP654}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP655}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP656}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP657}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP658}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP659}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP660}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP661}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP662}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP663}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP664}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP665}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP666}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP667}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP668}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP669}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP670}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP671}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP672}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP673}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP674}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP675}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP676}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP677}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP678}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP679}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP680}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP681}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP682}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP683}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP684}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP685}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP686}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP687}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP688}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP689}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP690}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP691}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP692}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP693}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP694}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP695}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP696}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP697}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP698}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP699}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP700}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP701}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP702}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP703}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP704}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP705}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP706}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP707}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP708}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP709}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP710}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP711}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP712}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP713}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP714}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP715}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP716}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP717}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP718}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP719}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP720}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP721}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP722}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP723}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP724}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP725}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP726}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP727}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP728}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP729}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP730}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP731}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP732}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP733}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP734}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP735}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP736}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP737}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP738}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP739}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP740}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP741}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP742}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP743}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP744}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP745}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP746}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP747}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP748}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP749}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP750}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP751}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP752}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP753}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP754}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP755}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP756}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP757}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP758}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP759}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP760}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP761}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP762}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP763}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP764}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP765}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP766}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP767}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP768}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP769}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP770}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP771}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP772}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP773}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP774}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP775}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP776}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP777}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP778}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP779}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP780}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP781}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP782}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP783}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP784}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP785}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP786}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP787}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP788}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP789}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP790}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP791}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP792}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP793}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP794}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP795}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP796}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP797}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP798}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP799}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP800}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP801}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP802}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP803}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP804}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP805}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP806}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP807}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP808}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP809}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP810}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP811}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP812}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP813}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP814}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP815}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP816}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP817}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP818}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP819}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP820}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP821}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP822}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP823}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP824}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP825}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP826}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP827}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP828}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP829}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP830}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP831}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP832}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP833}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP834}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP835}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP836}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP837}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP838}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP839}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP840}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP841}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP842}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP843}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP844}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP845}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP846}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP847}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP848}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP849}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP850}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP851}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP852}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP853}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP854}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP855}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP856}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP857}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP858}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP859}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP860}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP861}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP862}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP863}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP864}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP865}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP866}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP867}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP868}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP869}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP870}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP871}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP872}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP873}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP874}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP875}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP876}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP877}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP878}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP879}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP880}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP881}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP882}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP883}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP884}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP885}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP886}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP887}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP888}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP889}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP890}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP891}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP892}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP893}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP894}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP895}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP896}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP897}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP898}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP899}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP900}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP901}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP902}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP903}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP904}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP905}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP906}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP907}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP908}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP909}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP910}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP911}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP912}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP913}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP914}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP915}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP916}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP917}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP918}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP919}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP920}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP921}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP922}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP923}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP924}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP925}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP926}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP927}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP928}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP929}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP930}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP931}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP932}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP933}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP934}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP935}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP936}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP937}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP938}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP939}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP940}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP941}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP942}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP943}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP944}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP945}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP946}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP947}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP948}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP949}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP950}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP951}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP952}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP953}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP954}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP955}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP956}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP957}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP958}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP959}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP960}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP961}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP962}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP963}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP964}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP965}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP966}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP967}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP968}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP969}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP970}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP971}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP972}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP973}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP974}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP975}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP976}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP977}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP978}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP979}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP980}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP981}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP982}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP983}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP984}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP985}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP986}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP987}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP988}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP989}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP990}}{a factor with levels \code{C} \code{T}}
    \item{\code{SNP991}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP992}}{a factor with levels \code{G} \code{T}}
    \item{\code{SNP993}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP994}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP995}}{a factor with levels \code{A} \code{C}}
    \item{\code{SNP996}}{a factor with levels \code{C} \code{G}}
    \item{\code{SNP997}}{a factor with levels \code{A} \code{G}}
    \item{\code{SNP998}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP999}}{a factor with levels \code{A} \code{T}}
    \item{\code{SNP1000}}{a factor with levels \code{A} \code{C}}
  }
}
\examples{
data(vignette_kinship_dat)
}
\keyword{datasets}
