% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BICOSS.R
\docType{data}
\name{Y}
\alias{Y}
\title{A. Thaliana Simulated Phenotype matrix}
\format{
## `Y`
A matrix with 328 rows corresponding to the 328 ecotypes.
}
\usage{
Y
}
\description{
This is a phenotype matrix simulated from the 9,000 SNPs. SNPs at positions 450, 1350, 2250, 3150, 4050,
4950, 5850, 6750, 7650, and 8550 have nonzero coefficients. Further, the data was simulated under the linear mixed model
specified in the vignette and the BICOSS manuscript using the kinship matrix (kinship).
}
\keyword{datasets}
